[HumanBenchmark][Backend][Exercises] Refactor Exercises
